pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autodevos"
version = "0.1.0"
description = "AutoDevOS - OpenEnv Environment for Website Generation Tasks"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Zzeelan7", email = "zzeelan7@gmail.com"}
]
license = {text = "MIT"}

dependencies = [
    "openenv-core>=1.0.0",
    "pydantic>=2.0",
    "openai>=1.0.0",
    "aiohttp>=3.8.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]

[project.scripts]
server = "server.app:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[tool.setuptools]
packages = ["backend"]

[tool.openenv]
name = "autodevos"
description = "Website generation environment for RL training"
version = "0.1.0"
tasks = ["simple_landing_page", "portfolio_website", "responsive_ecommerce"]
docker_base = "python:3.11-slim"
enable_ui = true
ui_port = 7860
